Fix integration.
[git-p4: depot-paths = "//depot/qt/main/": change = 96134]

+++ src/codecs/qutfcodec.cpp
@@ -141,7 +141,7 @@ int QUtf8Codec::heuristicContentMatch(const char* chars, int len) const
 
 
 class QUtf8Decoder : public QTextDecoder {
-    ushort uc;
+    uint uc;
     int need;
     bool headerDone;
 public:
